Move database/user config building into start script.
I don't want to store all this stuff in Kubernetes. If I did, I would
have to update it all over the place in K8s to support changes in the
format pgbouncer wants, which would be annoying. The things it needs
(hosts, username, passwords) won't change. It's pretty ugly to do this
hashing stuff in bash, but it works.